+++ src/main/java/com/example/pharmacyback/service/ProductService.java
@@ -1,0 +1,22 @@
+package com.example.pharmacyback.services;
+
+import com.example.pharmacyback.repository.ProductRepository;
+import com.example.pharmacyback.model.Product;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ProductService {
+    @Autowired
+    ProductRepository productRepository;
+
+    List<Product> getAllProducts() {
+        try {
+            return productRepository.findAll();
+        } catch (Exception e) {
+            return null;
+        }
+    }
+}
